refactor: move useInput to its own folder.
