+++ docs/data/base/components/input/UseInput.js
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { useInput } from '@mui/base';
+import { useInput } from '@mui/base/useInput';
 import { styled } from '@mui/system';
 import { unstable_useForkRef as useForkRef } from '@mui/utils';
