+++ docs/data/base/components/input/input-pt.md
@@ -88,7 +88,7 @@ Note that `componentsProps` slot names are written in lowercase (`root`) while `
 ## Hook
 
 ```js
-import { useInput } from '@mui/base/InputUnstyled';
+import { useInput } from '@mui/base/useInput';
 ```
 
 The `useInput` hook lets you apply the functionality of `InputUnstyled` to a fully custom component. It returns props to be placed on the custom component, along with fields representing the component's internal state.
