+++ packages/mui-base/src/InputUnstyled/index.ts
@@ -2,9 +2,5 @@ export { default } from './InputUnstyled';
 
 export * from './InputUnstyled.types';
 
-export { default as useInput } from './useInput';
-
-export * from './useInput.types';
-
 export { default as inputUnstyledClasses } from './inputUnstyledClasses';
 export * from './inputUnstyledClasses';
