+++ packages/mui-base/src/useInput/index.ts
@@ -1,0 +1,3 @@
+export { default as useInput } from './useInput';
+
+export * from './useInput.types';
