+++ packages/mui-joy/src/Input/useForwardedInput.ts
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { useInput } from '@mui/base/InputUnstyled';
+import { useInput } from '@mui/base/useInput';
 import FormControlContext, { FormControlContextValue } from '../FormControl/FormControlContext';
 
 export default function useForwardedInput<Output>(
